backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
celery>=5.2.0
redis>=4.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0
httpx>=0.23.0
openai>=0.27.0
clerk-sdk-python>=0.1.0
python-jose>=3.3.0
passlib>=1.7.4
authlib>=1.2.1
python-multipart
bcrypt==4.0.1
pydantic[email]
aiofiles>=23.0.0